"check_style_score -> 2 of 4 passed. style score: 0/1 "

+++ m14/14.1 Dynamic Median Finding/Solution.java
@@ -95,7 +95,7 @@ public class Solution
         if(l==1)
  	   {
  		   
- 	 System.out.println(a[0]);
+ 	 System.out.println((double)a[0]);
  	   }
